requirements.txt
python-dotenv>=1.0.1

# LangChain / LangGraph
langgraph>=0.2.0
langchain>=0.2.14
langchain-openai>=0.1.22
openai>=1.44.0

# LiveKit Agents + plugins
livekit-agents>=0.11.0
livekit-plugins-openai>=0.7.0
livekit-plugins-elevenlabs>=0.3.0
livekit-plugins-silero>=0.2.0
livekit-plugins-turn-detector>=0.1.0
livekit-plugins-noise-cancellation>=0.1.0

# (opcional) si usas SDK RTC directamente en otros scripts
livekit>=0.21.0
